# docs/archive/Agentpay_snapshot_2026-03-16/backups/sensitive_backup_20260312_073532/openapi.yaml
openapi: 3.1.0
info:
  title: AgentPay API
  description: |
    **The Financial OS for AI Agents** — Universal payment gateway with AgentRank trust scoring,
    A2A escrow, multi-protocol support (x402, ACP, AP2, Solana Pay, Stripe), and KYA identity.

    ## Authentication
    All protected endpoints require an API key via:
    - `Authorization: Bearer <api_key>` header, or
    - `X-Api-Key: <api_key>` header

    ## Protocols Supported
    - **x402** — HTTP 402 paywall standard
    - **ACP** — Agent Communication Protocol
    - **AP2** — Agent Payment Protocol v2
    - **Solana Pay** — Solana USDC native
    - **Stripe** — Fiat card / bank payments

    Get your API key at [https://dashboard.agentpay.gg](https://dashboard.agentpay.gg)
  version: 1.0.0
  contact:
    name: AgentPay Support
    url: https://docs.agentpay.gg
    email: support@agentpay.gg
  license:
    name: MIT
    url: https://github.com/Rumblingb/Agentpay/blob/main/LICENSE
servers:
  - url: https://api.agentpay.gg
    description: Production
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /api/escrow/create:
    post:
      tags: [Escrow]
      summary: Create escrow
      description: Lock funds in escrow for a task
      operationId: createEscrow
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EscrowCreateRequest'
      responses:
        '201':
          description: Escrow created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EscrowResponse'
  /api/escrow/{escrowId}/approve:
    post:
      tags: [Escrow]
      summary: Approve escrow
      description: Approve work and release escrowed funds to payee
      operationId: approveEscrow
      security:
        - ApiKeyAuth: []
      parameters:
        - name: escrowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Escrow approved, funds released
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EscrowResponse'
  /api/escrow/{escrowId}/dispute:
    post:
      tags: [Escrow]
      summary: Dispute escrow
      description: Raise a dispute on an escrow transaction
      operationId: disputeEscrow
      security:
        - ApiKeyAuth: []
      parameters:
        - name: escrowId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for the dispute
              required: [reason]
      responses:
        '200':
          description: Dispute raised
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EscrowResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Format: 'Bearer <api_key>' or just '<api_key>'"
...
  schemas:
...
    EscrowCreateRequest:
      type: object
      required: [payeeAgentId, amount, taskDescription]
      properties:
        payeeAgentId:
          type: string
        amount:
          type: number
          description: Amount in USD
        taskDescription:
          type: string
        timeoutHours:
          type: integer
          default: 72
    EscrowResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [pending, locked, completed, disputed, resolved]
        amount:
          type: number
        payeeAgentId:
          type: string
        createdAt:
          type: string
          format: date-time
...